README.md: clarify meaning of coefficients
Since many users will interact with this package in the REPL, I thought it might be helpful to give a "visual" example of how to extract the various derivatives. At least to me, this was more immediate than interpreting `println("f''(t0) = ", eps1eps2(f(t0)))`. A danger is that without reading the paper, people might naively gravitate towards thinking the coefficient of `ϵ2` is the second derivative.